refactor browsers
